requirements.txt
tensorflow_datasets
Pillow